requirements.txt
# Core FastAPI and Vercel dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database and external service connections
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1

# Elasticsearch integration
elasticsearch==8.11.0

# Kafka integration
kafka-python==2.0.2
confluent-kafka==2.3.0

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# HTTP client for external API calls
httpx==0.25.2
aiohttp==3.9.1

# Data processing and ML (compatible with Python 3.12)
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Logging and monitoring
structlog==23.2.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1
bandit==1.7.5
safety==2.3.5
radon==6.0.1
